Fix OVERFLOW_BEFORE_WIDEN coverity warnings
*** CID 899134:         Integer handling issues  (OVERFLOW_BEFORE_WIDEN)
/lib/util/xdpsock.c: 1744             in xsk_ctx__create()
1738     	if (bufs == MAP_FAILED) {
1739     		pr_warn("ERROR: mmap failed\n");
1740     		goto err;
1741     	}
1742
1743     	/* Create sockets... */
>>>     CID 899134:         Integer handling issues  (OVERFLOW_BEFORE_WIDEN)
>>>     Potentially overflowing expression "4096U * opt->frame_size" with type "unsigned int" (32
>>> bits, unsigned) is evaluated using 32-bit arithmetic, and then used in a context that expects
>>> an expression of type "__u64" (64 bits, unsigned).
1744     	umem = xsk_configure_umem(bufs, NUM_FRAMES * opt->frame_size,
1745     				  opt->frame_size, umem_flags);
1746     	if (IS_ERR(umem)) {
1747     		ret = PTR_ERR(umem);
1748     		umem = NULL;
1749     		goto err;

*** CID 899132:         Integer handling issues  (OVERFLOW_BEFORE_WIDEN)
/lib/util/xdpsock.c: 835             in csum_tcpudp_nofold()
829
830     	s += (__u32)saddr;
831     	s += (__u32)daddr;
832     #ifdef __BIG_ENDIAN__
833     	s += proto + len;
834     #else
>>>     CID 899132:         Integer handling issues  (OVERFLOW_BEFORE_WIDEN)
>>>     Potentially overflowing expression "proto + len << 8" with type "unsigned int" (32 bits,
>>> unsigned) is evaluated using 32-bit arithmetic, and then used in a context that expects an
>>> expression of type "unsigned long long" (64 bits, unsigned).
835     	s += (proto + len) << 8;
836     #endif
837     	return (__wsum)from64to32(s);
838     }
839
840     /*

*** CID 899131:         Integer handling issues  (OVERFLOW_BEFORE_WIDEN)
/lib/util/xdpsock.c: 969             in gen_eth_frames()
963     {
964     	__u32 copy_len = frame_size;
965     	__u32 len = 0;
966     	unsigned int i;
967
968     	for (i = 0; i < NUM_FRAMES; i++) {
>>>     CID 899131:         Integer handling issues  (OVERFLOW_BEFORE_WIDEN)
>>>     Potentially overflowing expression "i * frame_size" with type "unsigned int" (32 bits,
>>> unsigned) is evaluated using 32-bit arithmetic, and then used in a context that expects an
>>> expression of type "__u64" (64 bits, unsigned).
969     		__u64 addr = i * frame_size;
970
971     		if (!len) {
972     			len = PKT_SIZE(&ctx->opt);
973     			copy_len = frame_size;
974     		}

+++ lib/util/xdpsock.c
@@ -68,7 +68,7 @@
 #define SO_BUSY_POLL_BUDGET     70
 #endif
 
-#define NUM_FRAMES (4 * 1024)
+#define NUM_FRAMES (4 * 1024UL)
 #define IS_EOP_DESC(options) (!((options) & XDP_PKT_CONTD))
 
 #define DEBUG_HEXDUMP 0
@@ -832,7 +832,7 @@ __wsum csum_tcpudp_nofold(__be32 saddr, __be32 daddr,
 #ifdef __BIG_ENDIAN__
 	s += proto + len;
 #else
-	s += (proto + len) << 8;
+	s += ((unsigned long long)proto + len) << 8;
 #endif
 	return (__wsum)from64to32(s);
 }
@@ -966,7 +966,7 @@ static void gen_eth_frames(struct xsk_ctx *ctx, struct xsk_umem_info *umem, __u3
 	unsigned int i;
 
 	for (i = 0; i < NUM_FRAMES; i++) {
-		__u64 addr = i * frame_size;
+		__u64 addr = (__u64)i * frame_size;
 
 		if (!len) {
 			len = PKT_SIZE(&ctx->opt);
